Fix the MWL usage on the deck of the week view on the homepage.
The MWL code was not part of the custom JSON that the highlight
command uses.  Additionally, the javascript would not try to
initialize an MWL instance even if it was there.

This has all been fixed and a bit of tidying up on the index page
and the highlight command was done along the way.

This fixes issue #50