Sanitize Javascript-rendered raw text
Deck descriptions, comments, card reviews, etc. are stored in two forms.
The 'rawtext' ('rawdescription', etc.) form has a mix of user-provided
markdown and HTML.  The 'text' form is pre-rendered HTML.  When the HTML
is rendered, it is converted from Markdown and sanitized to remove
unwanted HTML entities, such as `<script>` tags.

The XSS risk comes from places in the code that render the unsanitized
'rawtext'. This is typically in the 'preview' pane of text boxes, but is
also in the 'view' for a private deck.

This commit sanitizes these cases with DOMPurify after Markdown
conversion.