Fixed a litany of deckbuilder search issues.
The deckbuilder searcher has been plagued with bugs for 9 years. Now,
deckbuilder search functionality should match the site search bar.

The bulk of this fix is to change structure of the search query. We've
been using forerunnerdb which shares significant syntax with mongodb.
Previously OR (|) searches for integer conditions were not handled at all
(with text queries implicitly handled by the $in operation). Now, the OR
queries are properly handled by wrapping them with an $or operator. To
handle multiple $or operator in the same query, the entire query is then
wrapped by an $and operator.
Example, for o:2|3, the query that is made:

Old:
{
    "cost": {
        "$eq": [
            2    <----- Incorrect, $eq does not accept an array of
                        values. Also, the | token was not parsed
                        correctly, and the 3 value was dropped.
        ]
    },
    ...
}

New:
{
    "$and": [
        {
            "$or": [
                {
                    "cost": {
                        "$eq": 2
                    }
                },
                {
                    "cost": {
                        "$eq": 3
                    }
                }
            ]
        }, ...
    ]
}

Also, a minor fix in the state machine to properly deliniate | tokens
(line 192 token text-without-quotes) - exclude the | token from string
match. This fixes a bug where x:text1|text2 would error out.